Terminamos toda la persistencia

+++ src/main/java/domain/viaje/publico/sentido/SentidoRecorrido.java
@@ -3,7 +3,7 @@ package domain.viaje.publico.sentido;
 import domain.viaje.publico.Parada;
 
 public interface SentidoRecorrido {
-    Integer getDistanciaProxParada(Parada parada);
+    Double getDistanciaProxParada(Parada parada);
 
     boolean isBetween(int indInicial, int indFinal, int indexOf);
 }
